Fix "Jump to Method Definition"
This command fails because of the missing "rubygems" library for the custom Ruby 1.8 interpreter of TextMate 2. The only thing its used for is finding the paths of all installed gems. This makes no sense nowadays anyway, as every supported Rails version requires Bundler, which in turn can list only the gems used in a Rails project.

Shelling out and using `bundle show --paths` restores the functionality of this command, adds no additional dependencies and, needing to search fewer directories, is quicker too.

Fixes #23